Fix nullptr defined as 0 on g++.
SafeInt3.hpp used an invalid check to determine wether nullptr should be defined as 0 with g++ compiler. Checking that we use c++11 or above fixed the issue.
However, using a valid check broke the compilation because boost::function cannot be resetted by assigning a nullptr to it. Using default constructed boost::function for reset fixed this.