Fix header reading on linux listener using HTTPS.
The first asynchronous read of the stream was stopping at the first CRLF (after the HTTP version) instead of reading until CRLFCRLF (the end of the header).

This bug was not always noticeable because, as stated in the boost documentation:
"After a successful async_read_until operation, the streambuf may contain additional data beyond the delimiter. An application will typically leave that data in the streambuf for a subsequent async_read_until operation to examine."
Therefore the bug appeared only when trying to read the body of the request and the streambuf filled by async_read_until did not contain the complete header.