Add support for Android NDK unified header files for OpenSSL.
This patch updates the OpenSSL Makefile and patches used for
Android builds so that it supports the unified headers and sysroot
found in newer Android NDKs since r14. This is required for Android
NDK r16 and later as the platform specific headers have been
removed. Additionally, the default prefix (install location) has
been changed in the Makefile to encode more variant information,
however the configure.sh script has been modified to override these
with the older defaults for the time being.