Improved the cpprestsdk build process on Android.
The Android configure.sh shell script and CMake build files have
been updated for cpprestsdk on Android with the following changes:

- Added new --api and --stl command line switches to the
  Android_build/configure.sh shell script that control the Android
  API level and C++ runtime to build against.
- Use the base CMake toolchain configuration files that come with
  Android NDK r13b and later if they are available (requires CMake
  3.6 or later to be installed on host system), otherwise fallback
  fallback to using the older third-party android-cmake.
- Fix a few compilation errors when building with libc++ on Android,
  namely to do with assumptions of using libstdc++ paired with
  checked the ANDROID or __ANDROID__ preprocessor symbols.
- Make use of the new features added in subsequent PRs for
  building OpenSSL and Boost for Android.

Note that cpprestsdk seems to no longer build with libstdc++ 4.8
or 4.9 library headers that ship with the Android NDKs due to the
use of std::to_string introduced in commit
639726178046803d7193785b2eb2e75137df40f3 so the default C++ runtime
has been changed to c++_static in configure.sh for the time being.
This problem can probably be fixed in a separate patch, as its
outside of the scope of this one.