Show found version of openssl (cmake UNIX)
This is an addition to commit cb7ca74. A message is added for cmake
UNIX target systems, which shows the version of openssl found by
pkg_search_module(). Furthermore, the minimum required version of 1.0.0
is added to pkg_search_module() like in the subsequent
find_package(OpenSSL 1.0.0 REQUIRED).

Example of the cmake output including the message:
  -- Found PkgConfig: /usr/bin/pkg-config (found version "1.6.3")
  -- Checking for one of the modules 'openssl>=1.0.0'
  -- Found openssl version 1.1.1d on system
  -- Performing Test _SSL_LEAK_SUPPRESS_AVAILABLE